+++ src/ArduinoJson/Deserialization/Readers/FlashReader.hpp
@@ -4,16 +4,14 @@
 
 #pragma once
 
-#include <ArduinoJson/Namespace.hpp>
-
-#if ARDUINOJSON_ENABLE_PROGMEM
-
 namespace ARDUINOJSON_NAMESPACE {
-class UnsafeFlashStringReader {
+
+template <>
+struct Reader<const __FlashStringHelper*, void> {
   const char* _ptr;
 
  public:
-  explicit UnsafeFlashStringReader(const __FlashStringHelper* ptr)
+  explicit Reader(const __FlashStringHelper* ptr)
       : _ptr(reinterpret_cast<const char*>(ptr)) {}
 
   int read() {
@@ -27,12 +25,13 @@ class UnsafeFlashStringReader {
   }
 };
 
-class SafeFlashStringReader {
+template <>
+struct BoundedReader<const __FlashStringHelper*, void> {
   const char* _ptr;
   const char* _end;
 
  public:
-  explicit SafeFlashStringReader(const __FlashStringHelper* ptr, size_t size)
+  explicit BoundedReader(const __FlashStringHelper* ptr, size_t size)
       : _ptr(reinterpret_cast<const char*>(ptr)), _end(_ptr + size) {}
 
   int read() {
@@ -50,15 +49,4 @@ class SafeFlashStringReader {
     return length;
   }
 };
-
-inline UnsafeFlashStringReader makeReader(const __FlashStringHelper* input) {
-  return UnsafeFlashStringReader(input);
-}
-
-inline SafeFlashStringReader makeReader(const __FlashStringHelper* input,
-                                        size_t size) {
-  return SafeFlashStringReader(input, size);
-}
 }  // namespace ARDUINOJSON_NAMESPACE
-
-#endif