+++ CHANGELOG.md
@@ -4,7 +4,7 @@ ArduinoJson: change log
 HEAD
 ----
 
-* Added support for custom writer classes (issue #1088)
+* Added support for custom writer/reader classes (issue #1088)
 * Added conversion from `JsonArray` and `JsonObject` to `bool`, to be consistent with `JsonVariant`
 * Fixed `deserializeJson()` when input contains duplicate keys (issue #1095)
 * Improved `deserializeMsgPack()` speed by reading several bytes at once